Move weaveworks/mesh into a repackaged location
This allows us to make changes to it locally.
