+++ protokube/pkg/gossip/mesh/glogger.go
@@ -18,7 +18,7 @@ package mesh
 
 import (
 	"github.com/golang/glog"
-	"github.com/weaveworks/mesh"
+	"k8s.io/kops/protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh"
 )
 
 // glogLogger sends mesh log messages to glog
