+++ protokube/pkg/gossip/mesh/peer.go
@@ -18,8 +18,8 @@ package mesh
 
 import (
 	"github.com/golang/glog"
-	"github.com/weaveworks/mesh"
 	"k8s.io/kops/protokube/pkg/gossip"
+	"k8s.io/kops/protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh"
 )
 
 // Peer encapsulates state and implements mesh.Gossiper.
