+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/BUILD.bazel
@@ -21,7 +21,11 @@ go_library(
         "surrogate_gossiper.go",
         "token_bucket.go",
     ],
-    importpath = "github.com/weaveworks/mesh",
+    importpath = "k8s.io/kops/protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh",
+    tags = [
+        "peer_name_alternative",
+        "peer_name_hash",
+    ],
     visibility = ["//visibility:public"],
     deps = [
         "//vendor/golang.org/x/crypto/nacl/box:go_default_library",
