+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/circle.yml
