+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/connection.go
