+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/connection_maker.go
