+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/gossip.go
