+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/gossip_channel.go
