+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/local_peer.go
