+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/logger.go
