+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/peer_name_mac.go
