+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/protocol_crypto.go
