+++ protokube/pkg/gossip/mesh/repackaged/weaveworks/mesh/surrogate_gossiper.go
